Moved CommandHandler logic into the Bot class.
